requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
loguru==0.7.2
python-dotenv==1.0.1
pyjwt==2.9.0
loguru==0.7.2
httpx==0.27.2
fastapi-sessions==0.3.2

